@@ -1,11 +1,11 @@
 {
-  "name": "@maaxgr/ang-jsoneditor",
-  "version": "13.0.5",
+  "name": "ng-jsoneditor2",
+  "version": "14.0.1",
   "license": "MIT",
-  "homepage": "https://github.com/MaaxGr/ang-jsoneditor",
+  "homepage": "https://github.com/juicycleff/ang-jsoneditor",
   "repository": {
     "type": "git",
-    "url": "https://github.com/MaaxGr/ang-jsoneditor"
+    "url": "https://github.com/juicycleff/ang-jsoneditor"
   },
   "author": "Max Großmann <github@maax.gr>, Mario Mol <mariohmol@gmail.com>, Manish Sodavadiya <manish_it56@yahoo.com>",
   "keywords": [
@@ -16,8 +16,8 @@
     "angular"
   ],
   "peerDependencies": {
-    "@angular/common": "^13.3.0",
-    "@angular/core": "^13.3.0",
+    "@angular/common": "^14.2.12",
+    "@angular/core": "^14.2.12",
     "jsoneditor": "^9.1.4"
   },
   "dependencies": {
